separate basic data structures from mrcProfiler.cpp
